Parametrize service_host value in the test config
Parametrize service_host value in the test config to be based on
actual value of SERVICE_HOST from devstack/vmgate config. This is
needed for the vm-gate integration piece. The vm-gate driver to devstack
uses 127.0.0.1 instead of localhost which causes a mismatch in the
reddwarf URLs checked by the tests, so these need to be set correctly.

Fixes bug 1103778

Change-Id: If544db2eb0e9fec5c5b35f6551ae0e7b648d35ae