Fixed redstack to enable host ips other than 10.0.0.1
There were a few places in redstack, reddwarf conf files, and diskimage-builder
files where we were assuming that the host would always be at 10.0.0.1. This was
tripping up the devstack-vm-gate which has the host as 10.1.0.1 . Fixed the
scripts and config files to use the NETWORK_GATEWAY from the devstack configuration.
We fall back to 10.0.0.1 if no NETWORK_GATEWAY is specified as the default behaviour.

This fixes bug 1108372

Change-Id: Ib53ff3548269abd1b9e006d71382a417237f0068