Install PyPi dependencies alongside apt dependencies for dib elements
The python-novaclient package in apt is really old (2012.1-0ubuntu1) and
has a pinned dependency on prettytable 0.5. All of the newer clients have
a dependency on prettytable 0.6 - 0.8. This is causing issues on the guest
agent.

This updates the dib elements so that dependencies from PyPi are installed
alongside apt dependencies.

Fixes bug 1189641

Change-Id: I5233aad2045cbec8610569c40e50bf26defe2dfa