Custom Guest Log Directory Causes Init Failure
The log_dir in trove-guestagent.conf.sample currently defaults to
'/tmp', and the log_file to 'logfile.txt'.

If the log_dir value is changed to, say, '/var/log/trove', guest
initialization will fail because the directory does not exist. Even
if log_dir is set to '/var/log', initialization will fail due to
insufficient write permissions in said directory.

Solution: The log_dir specified in trove-guestagent.conf.sample
should be created on-demand if necessary.

Change-Id: Ic819aaa08f8a78483e1c0ae0aa8af370c6705562
Closes-Bug: #1219137