Adds support to kickstart diff datastore versions
Just like PACKAGES string can be overridden by supplying value via
defining the environment variable prior to running redstack kick-start,
the same facility is now extended to DATASTORE_VERSION string as well.

If the DATASTORE_VERSION is not defined as env var, then the behavior
is the same that would happen without this commit. This commit does
not  alter the default behavior in any way.

USAGE: export DATASTORE_VERSION="-2.4.9"; export PACKAGES=" "; \
       redstack kick-start mongodb

The reason to set PACKAGES=" ", is to prevent the triggering
of Package update at guestagent prepare() time if you are
already prebaking the required version you are interested in,
into the dib image itself. The PACKAGES update at guestagent
prepare() time is partially broken and you may want to avoid
that path to get the desired version of your datastore installed
and instead choose this VERSION string export mechanism

This functionality allows 3 things:
1) Specify the elements for your specific datastore version
   (:eg mongodb 2.4.9) to be picked by dib from an alternative
   location such as for example for mongodb, from
   scripts/files/elements/ubuntu-mongodb-2.4.9 rather than
   the default scripts/files/elements/ubuntui-mongodb when
   no DATASTORE_VERSION is specified. This allows you for
   example to define a different install.d/* element for your
   specific datastore version if it is different from default

2) when redstack invokes trove-manage to register the
   glance image with specific version in trove DB it will
   set the version for eg as 'mongodb-2.4.9'

3) Also if you are taking advantage of dib tools or redstack
   kick-start in trove-integration to build your internal
   company specific image/version of your datastore and
   want to point these tools at your custom elements that
   might be installing packages from custom (non distro
   standard ) repos, then this change allows you to do
   so without impacting the standard behavior and
   conventions for the opensource trove-integration project
   (convention is to support only standard distro repos
    of datastore packages officially/publicly)

Closes-Bug: #1279098
Change-Id: I56d3f6d5706731ab3b499e2629b27dd55b1a5f04