Reset the no_proxy when the node is installed behind a proxy
commit: Iccc4354d8eae0256acd63a5f12232d50d1ec4049
Date: Tue May 27 14:20:00 2014 -0500
Implements: bugfix BUGFIX
Closes-Bug: 1322200/1291111
Reason of the change: The installation will be failure if the network
proxy is set. The reason of the failure is that any REST request sent
to the installed OpenStack itself will be redirected to the proxy
server. By default, the no_proxy is set to localhost only, which needs
to be reset to the localhost + HOST_IP + FLOATING_RANG + FIXED_RANGE.
Note: This fix works with another fix in Devstack
(Change-Id: I3cdace056b74fec111d24d222a886716a063d340)
Changes:
 Add "export USERHOME=$HOME" in redstack
     This is used by devstack installation to save the new no_proxy
in the home direction

Change-Id: Iccc4354d8eae0256acd63a5f12232d50d1ec4049