commit: Iccc4354d8eae0256acd63a5f12232d50d1ec4049 Author: Zu Qiang Date: Thu May 22 11:35:00 2014 -0500
Implements: bugfix BUGFIX
Closes-Bug: 1322200/1291111

Reason of the change: The installation will be failure if the network 
proxy is set. The reason of the failure is that any REST request sent 
to the installed OpenStack itself will be redirected to the proxy 
server. By default, the no_proxy is set to localhost only, which needs 
to be reset to the localhost + HOST_IP + FLOATING_RANG + FIXED_RANGE.

Note: This fix works with another fix in Devstack 
(change-Id Change-Id: I6b8e2ba2b7e78fa280283d5ee82d618d4f44ec5a)

Changes:

 Add "export USERHOME=$HOME" in redstack
     This is used by devstack installation to save the new no_proxy 
in the home direction

 Add "Reset no proxy" in redstack
     This is used to reset the no_proxy, once the devstack installation
is done.
            #Reset no proxy
            if [ -f $USERHOME/noproxy ]; then
                  . $USERHOME/noproxy
            fi


Change-Id: Iccc4354d8eae0256acd63a5f12232d50d1ec4049