Use 'sudo install' for conf files instead of 'cp'
In a devstack-gate environment, the devstack user (stack) is used to run
devstack, and so the /etc/trove directory is writable only by the stack
user. This causes kick-start to fail when the ubuntu user tries to copy
the trove-guestagent.conf to the /etc/trove CONF directory.

Fixed this by using "sudo install" instead of a plain "cp" to copy the
config files over to the /etc/trove directory.

Closes-bug: 1333454
Change-Id: Ie2df541b27ceb44d1ec685a313fed19a1f1ca40e