Associate flavor types with datastore versions
This feature provides the ability to associate flavor types with datastore versions.
The management api will provide the ability to add or delete this information.
The normal user can view the flavor list for the default version or a
specified datastore version id.The redstack ensures the mapping of all default flavors to the
default version takes place at the time of deployment.

Implements: blueprint associate-flavors-datastores
Change-Id: I726e76f5dbe23edb816c64b1097ed3eebc4f6f5b