Move elements from ubuntu-guest to mysql and percona
The files 20-apparmor-mysql-local and 50-percona-server which were in
the ubuntu-guest element are only relevant for mysql and percona guest
agents and therefore having them in ubuntu-guest is not ideal. While
not catastrophic, it does mean that these settings show up in all
guest images.

The files have been relocated to ubuntu-mysql and ubuntu-percona and
renamed to keep the right sequence of execution and an update of
repositories is forced after adding the percona repositories because
this is now occuring in the install.d phase of mysql (earlier it was
happening in pre-install of ubuntu-guest so the forced update wasn't
required).

Change-Id: I6e8fd0fde37ee9689b9f6a669a33d579b2ba6c64
Closes-Bug: #1349553