Move elements from ubuntu-guest to mysql and percona
The files 20-apparmor-mysql-local and 50-percona-server which were in
the ubuntu-guest element are only relevant for mysql and percona guest
agents and therefore having them in ubuntu-guest is not ideal. While
not catastrophic, it does mean that these settings show up in all
guest images.

The files have been relocated to ubuntu-mysql and ubuntu-percona. The
file that adds the percona key has been put in a newly created
pre-install.d directory and renamed. The other (apparmor) has
similarly been moved into this new pre-install.d directory.

Change-Id: I6e8fd0fde37ee9689b9f6a669a33d579b2ba6c64
Closes-Bug: #1349553