Fix malformed pip-install lines
The pip install command attempted to provide version hint requirements
on the command line but failed to quote the individual package
names. This resulted in the creation of files in the "/" directory of
the guest with names corresponding to the versions specified. The
correct way to specify version restrictions is to quote the package
names.

Change-Id: Ida186454d247d827312fd004d84efba13f780b0c
Closes-Bug: #1406581
Closes-Bug: #1409053