Fix malformed pip-install lines
The pip install command attempted to provide version hint requirements
on the command line but failed to quote the individual package
names. This resulted in the creation of files in the "/" directory of
the guest with names corresponding to the versions specified. To make
it easier to maintain in the future, a requirements.txt file for
ubuntu and fedora and a extra-data.d step which will copy this from
the source tree into the guest image (extra-data.d/15-reddwarf-dep)
have been added, and the 15-reddwarf-dep step in install.d can
reference the requirements.txt file.

requirements.txt has been updated to reflect the versions of packages
that trove has in its requirements.txt. Finally the dependency issue
with testtools and unittest2 have been addressed by adding them to the
requirements.txt file and adding the --upgrade option to pip install.

Change-Id: Ida186454d247d827312fd004d84efba13f780b0c
Closes-Bug: #1406581
Closes-Bug: #1409053