readme clarifications for new developers using redstack
There are 3 areas in the README.md that could benefit from clarification
for new developers:

1. change the recommended development environment from Ubuntu 12.04 (precise)
to 14.04 (trusty). Seems like this is the current viable environment in use by developers.
2. expanded details of navigating the multiple screens set up by redstack
3. clarify that kick-start is necessary for a usable default test.conf file
(in the absence of a developer override of TROVE_CONF_DIR)
i.e., cmd_build_and_upload_image->cmd_set_datastore

Closes-Bug: #1409038
Change-Id: I129e9f24aca7ad401724b6c3bdb8408879eb026c