kick-start doesn't warn on stale image file
Once you run './redstack kick-start <datastore>' the image is cached.
If you then pull a more recent copy of the trove-integration repo and
run kick-start again, it will use the cached image.  It will do this
even if the corresponding elements have changed.

This fix checks all the files in the elements directory and prints
out a warning in red if the cached image directory is older than any
of the files that it is built with.  An option to force a rebuild of
the image was also added.

A few unsafe constructs in the code changed were also fixed.

Change-Id: I0922c92525ab3ea8678d6dd067ce8bff7d961c31
Co-Authored-By: Petr Malik <pmalik@tesora.com>
Closes-Bug: #1415690