Clean up *_TENANT variables in redstack
In the fix for using the admin tenant for performing flavor-create
(https://review.openstack.org/#/c/155151), the variable SERVICE_TENANT
was left behind. Since it is no longer needed, it was removed.

An ADMIN_TENANT_NAME variable was created in its stead and used
throughout the code where 'admin' was hard-coded for the tenant name.

This should make the code inherently easier to read.

Change-Id: I728b2b48d2eca82d359387f83f686646a38f5c6d
Closes-Bug: #1421463