Refactor redstack to use local.conf
devstack moved away from a localrc file and towards using a
[[local|localrc]] section in the new local.conf file a while ago
(Jan 10, 2014-https://github.com/openstack-dev/devstack/commit/38e38fb )
Although there is no indication that support for the old-style file is
being withdrawn, it's probably time to upgrade the redstack script to use
the new format.

This becomes a more important issue when different configurations
are desired.  For example, in order to support profiling (osprofiler),
entries have to be made in the corresponding trove configuration files.
If we want the ability to turn profiling on and off, we need to switch
to the new local.conf file.  This would allow us to include
configurations such as the following:

[[post-config|$TROVE_CONF]]
[profiler]
enabled = True

where the line 'enabled = True' could be generated from redstack in
the form of 'enabled = $ENABLE_PROFILER'

Being able to add commands like that depends on
https://review.openstack.org/#/c/155486/ merging, although order is
not important as devstack will ignore configuration files that do not
exist.

Refactoring will also permit user specified configuration to become
more in-depth, allowing greater control over the trove/devstack
environment.  This was enhanced by including a USER_LOCAL_CONF
variable that defaults to $HOME/.local.conf
All the contents of that file are now included in devstack's local.conf
file on each invocation of 'redstack install.'

Backwards compatibility was maintained by deferring to an existing
localrc file in the devstack base directory.  In order to use the
new stype configuration file, the localrc file must first be
deleted or renamed.  This would be a good time to move any 'user'
configuration to $HOME/.local.conf

Stubs for the four Trove configuration files were added as well.

Change-Id: I319476b29aada7f24829ae2308bf88c8c4bdafa7
Closes-Bug: #1421859