Fix redstack issues (plugin use, stop and start)
Due to having multiple patchsets in flight at the same time
(https://review.openstack.org/#/c/158540/
https://review.openstack.org/#/c/167047 and
https://review.openstack.org/#/c/155913 ),
'start' and 'stop' still relied on having the Trove setup support in
devstack (the lib/trove file).

This needs to be changed to use the devstack/plugin.sh file in the
Trove repo once the devstack code is gone.  If not, when the devstack
patchset merges (https://review.openstack.org/#/c/167447 ) the
'stop' and 'start' commands will stop working.

Also made the using the plugin dependent on having the devstack
setup code gone. This fixes the issue of having devstack start
up two sets of Trove services.

Change-Id: Ic1e5a809a47b3f00d5eaec30237e37b73c676aad
Closes-Bug: #1448189