Allow specifying devstack branch in redstack
At present, redstack clones the most recent version of devstack.
Sometimes this is undesirable as devstack can at times become unstable.
A solution to this would be to add an environment variable or option to
redstack to specify the branch or commit of devstack to use (with the
default remaining 'master').

With this change you can now specify:

DEVSTACK_BRANCH=stable/kilo ./redstack install

or

DEVSTACK_BRANCH=7ef2464 ./redstack install

and redstack would switch to the corresponding branch/commit before
running './stack.sh'

Optionally, for persistence, it can be defined in either options.rc or
~/.redstack.options.rc and it will be picked up from there.

The repo is switched, even if the devstack directory already existed.

If an invalid branch/commit is specified, redstack stops with an error
message.

Change-Id: I50c64db917bb87bac5d1b1d4d0231270577e26a8
Closes-Bug: #1449627