Percona Xtradb Cluster image building scripts
Making a new datastore pxc image with the Percona Xtradb Cluster 5.5
packages from the public ubuntu repository.

Ran into some odd issues when booting the image where the guest agent
would start and stop the pxc service and about 60% of the time the
upstart process would start mysql when the guest was expecting mysql
to be stopped.

Implements: blueprint support-pxc-56
Change-Id: I0f95341e4e89572f784d87ff593939d1b464779c