Fix trove-functional-dsvm-gate tests
This bug fix forces the openstack client endpoint list, create
and delete commands to use the keystone v3 API instead of the
v2 API.

The move to keystone v3 in Devstack means that the endpoint-list
command of the openstack client no longer returns the list of all
v3 endpoints. For the dsvm gate tests to function we need to delete
the old swift endpoints and recreate new ones (reachable from the
guest).

Change-Id: I2ae8f54b257b9cf296bc29cf90990afe54576673