Revert "Fix trove-functional-dsvm-gate tests"
This reverts commit fd52298c7911f92eb8930fe1a8cff0f3798f2967

This assumes that the version of openstackclient being used
has the -c and -f options, which is not true in stable/kilo
where osc is capped at 1.0.4.

If this project is going to be branchless it should be
gated against stable compat jobs like Tempest is so we
don't regress stable branches by changes which only have
master in mind.

Change-Id: I3fd9726765c6b7e785a6f7d561918a5673add380
Closes-Bug: #1479358