Fix openstack endpoint list usage for stable branches
Commit fd52298c7911f92eb8930fe1a8cff0f3798f2967 did not take into
account capped versions of python-openstackclient so the stable branches
are broken with the gate-trove-functional-dsvm-mysql job.

This adds the old code back but makes it conditional to only stable/juno
and stable/kilo.

Other areas that need special attention (sourcing of files) was
also modified to not fail if on an older branch.

Co-Authored-By: Nikhil Manchanda <SlickNik@gmail.com>
Closes-Bug: #1479358

Change-Id: Iddeaaaf77b326945200be0e59df73745122b7f7b