Change redstack to only install server plugin
Using the Trove devstack plugin was dependent on having a
python-troveclient also 'enabled.' This had been changed so that
there is no need to separately enable a client plugin, so the code
to do so has been removed.

Variables have been added/modified to control how the plugin
is installed, and how the client is installed within the plugin.

Change-Id: Ic984dab5acc75934c5e9de19ac8bc4e2a6c4049a
Depends-On: If5a76360249827c19759a5c90dc00f70c59c32cc