Fix ubuntu image builds
The format for the SHA256SUMS file recently changed to refer to the
images using the release version nomenclature (e.g. ubuntu 14.04) as
opposed to the release code name (e.g. trusty). This was causing ubuntu
based image builds to fail. Fix this by explicitly specifying the release
version as the BASE_IMAGE_FILE so that dib can use this consistently
throughout.

Change-Id: I36bfba75c22b0a153eda2fcc3137fcefa9a29536