Fix gate (caused by devstack using clouds.yaml)
devstack recently changed upload_image to use the clouds.yaml file
instead of tokens.  This caused redstack to break in the gate as the
file is written by devstack in the user's home directory.  Redstack
changes the home directory, so the file isn't found.

Added a call to create the file for the gate job.

Change-Id: Ia541930a5cb99068434fc41ade92efc8c383455f
Closes-Bug: #1498626