Fix gate (caused by devstack using clouds.yaml)
devstack recently changed upload_image to use the clouds.yaml file
instead of tokens.  This caused redstack to break in the gate as the
file is written by devstack in the 'stack' user's home directory.

Added code to link the file from the 'stack' user's actual home directory
to the redstack user's 'home' directory.

Closes-Bug: #1498626
Co-Authored-By: Nikhil Manchanda <SlickNik@gmail.com>
Change-Id: Ida0aa155b1e1ec7b8e9fa73006d3db9a88ffea02