Fix gate (caused by devstack using clouds.yaml)
devstack recently changed upload_image to use the clouds.yaml file
instead of tokens.  This caused redstack to break in the gate as the
file is written by devstack in the user's home directory.  Redstack
changes the home directory, so the file isn't found.

Added code to link the file from the user's actual home directory if it
exists there and doesn't exist in redstack's 'home' directory.

Change-Id: Ida0aa155b1e1ec7b8e9fa73006d3db9a88ffea02
Closes-Bug: #1498626