Build Vertica UDx examples
Vertica comes with SDK example code that provides a number of User
Defined Functions and Loads. However they need to be built manually,
which isn't possible to do in a running instance. Therefore they should
be built by the elements so that they are available in the built images.

Change-Id: Icb7d9b1551124a700ad767eeba10be939ffa9aaa
Partial-Bug: bug/1531510