Allow redstack setup large volume and force qemu in devstack
Large volume is helpful for testing mongo where the volume should be
at least 5 gigs if the smallfiles parameter is not set by default.

Force qemu was useful when kvm was enable on my vm but causes the vm
to freeze as soon as a guest vm was booted on the instance. Found that
this was a kernel bug with nested kvm instances.

Change-Id: I89ac30243e3c47cbd36b0a0890242a2a8031bea4