fix the gate removing the heat-cfntools dependency
In the bug i describe 2 options we have to fix the gate right now and
this is option 1.

This is removing the heat-cfntools from the guest image we test
with. Since we do not test anything in the gate that requires heat
this helps minimize the elements we need as well as fix our issue with
pip 8.0 because we can control the packages installed on the vm.

Change-Id: I7069fcab3d16a960dc52757c087907ee8c41b309
Closes-Bug: #1536530