Elements incorrectly write entry to /etc/hosts file
When elements are executed, they are executed on some machine and in a
chroot environment. When the hostname command is executed, it returns
the hostname of the machine on which the image is being built.

The two elements being deleted here attempt to add an entry to the
/etc/hosts file for the loopback interface with the value of
'hostname'. Ostensibly the intent was to have the name of the host
where the guest was launched show up here. Well, that doesn't quite
happen.

Change-Id: Icb1fe9fe3fbe22dc38caebd93eba22a134df2878
Closes-Bug: #1529471