Use integer and string IDs for flavors
In a recent changeset (https://review.openstack.org/#/c/271096 )
the flavors created by redstack were changed to facilitate
testing of non-MySQL datastores.  Using string IDs was removed
from that change when it was discovered that it cause the kilo
tests to fail.

This changeset puts them back (with proper handling for kilo) and
consolidates all the creation in one spot.

The vertica.conf file was also fixed to use the proper resize flavor.

Change-Id: I326080f45a119c0e377498a7e9dfb1c8f39c5dbb