Add region support (for guest_log int-tests)
The guest_log int-tests create a swift client.  Since there
is no region code defined currently in test.conf a warning
is emmitted

"Providing attr without filter_value to get_urls() is
deprecated as of the 1.7.0 release and may be removed in the
2.0.0 release. Either both should be provided or neither
should be provided."

The region name is now taken from the environment
or defaulted to the devstack 'RegionOne' value.

Change-Id: I77981d282a6504379bb75a1e56cfb410e1c54f9a