Don't reset HOME in redstack
Currently redstack resets HOME as the first thing that it does. It
isn't particlularly apparent why this is required; it may have been
needed in a long ago time when paths to thing were
trove-integrations/scripts relative but they aren't now.

I've left the code in for now, with a #TODO just in case someone has
scripts that leverage or extend redstack and they want the old
behavior. I'll delete this code later. Hence this change has
Partial-Bug and not Closes-Bug.

Change-Id: If0f7329ea37f6d92f10b815c7a96eaee97d8b6d4
Partial-Bug: #1552048