Change redstack to use default config values
The plugin code has been enhanced to allow configuration
of Trove using variables. The default there is to use the
actual config values in the Trove repsoitory. This has
been mirrored here, in that the variables are listed so
that a user/developer can modify them if they wish, but
they do not override the default values anymore.

The proper place to add these variables would be in a
'local' options file, such as options.rc or
~/.redstack.options.rc

Change-Id: I38bc7190f1c112b4ae3cd6614f1d6cb625b3387d
Depends-On: I2c568fdfa05064682f372b996a5aebcdd4e93ef3