Trove image build fails getting percona key
This fails often in the gate and I ran into it migrating
trove-integration into trove-image-builder.

The most common cause for this is that keys.gnupg.net fails. While
there are multiple servers proxied for this service, sometimes they
fail. In general people have found keyserver.ubuntu.com to be more
reliable.

Change-Id: I97dcb9120c75129d8a0aaed9ddd39b9e5fa7e608
Closes-Bug: #1579094