Add 'fault' flavor to test persisting an error msg
In order to test that error messages are now persisted,
a method of guaranteeing that a Trove create command
will fail is needed.  The simplest way to do that is to
attempt to allocate more resources from Nova than it
has available.  To this effect, a 'fault' flavor has
been added.

Also bumped up the disk on mariadb to handle more
operating systems.

Change-Id: I32a5ed229988f047db44183c71f0462fa6f1990d