Add another 'fault' flavor
Apparently there are two code paths for a failure
creating a Nova instance. One returns immediatesly
and the other takes some time to fail.  The case
where the initial command to Nova succeeds only
to fail later needs another flavor to test it.

A new flavor with not enough disk to install the
Trove image has been created to test this scenarion.

Change-Id: I5d3339e9cbfd6aeb0c3ff6936fefa8dbe9e841f8