[WIP][DNM] upper-constraints.txt handling
Currently, trove-integration attempts to get an upper-constraints file
into the guest by looking at the tip of the appropriate branch. This
is fine for developer side testing but not ideal for the CI.

This change looks to see whether the CI setting (DEST) is set, and use
that to see whether $DEST/requirements/upper-constraints.txt is a
file, and if yes to both uses that. If not, it attempts to guess based
on the branch name.

This therefore covers both cases, the CI and the developer environment
and should allow for chaining with jobs that attempt to change u-c.

Co-Authored-By: Tony Breeds <tony@bakeyournoodle.com>
Co-Authored-By: Amrith Kumar <amrith@tesora.com>
Closes-Bug: #1610938
Change-Id: I06406051c849c924371f22b9fe1b03ab99e300f2