Enhance 'redstack clean' regarding modules
Adds a sql command to make sure the modules are not marked
as having been applied to an instance so that they can be
deleted (otherwise an error may occur if the instance hasn't
gone away yet).

Enhanced the logic to only sleep if objects are found
and to keep trying until the objects are gone (up to 10 times).
This makes runs with nothing to delete faster and negates having
to run the script multiple times if objects delete slower than
expected.

Fixed secgroup delete to look for the 'default' name instead
of an id of '1' (which could be incorrect) and added code
to delete server-groups as well.

Change-Id: If429b2d04111a6e8adbb90c001002215a927d1da