Enhance 'redstack clean' regarding modules
Adds a sql command to make sure the modules are not marked
as having been applied to an instance so that they can be
deleted (otherwise an error may occur if the instance hasn't
gone away yet).

Enhanced the logic to only sleep if objects are found
and to keep trying until the objects are gone (up to 10 times).
This makes runs with nothing to delete faster and negates having
to run the script multiple times if objects delete slower than
expected.

Added a check for server-groups to delete them if they exist.

Fixed security group delete to look for the 'default' name instead
of an id of '1' (which could be incorrect).  Also switched it to
use the OpenStack client, since Nova was reporting a deprecation
warning:

  WARNING: Command secgroup-delete is deprecated and will be removed
           after Nova 15.0.0 is released. Use python-neutronclient or
           python-openstackclient instead.

Unfortunately python-openstackclient has a warning of its own:

  WARNING: openstackclient.common.utils is deprecated and will be
           removed after Jun 2017. Please use osc_lib.utils

Hopefully this warning will go away once the 'openstack' client is
updated.

Change-Id: If429b2d04111a6e8adbb90c001002215a927d1da