Update the Signing Key for Percona Debian and Ubuntu Packages
Attempting to build images fails now with an error because of a change
in the signing key for Percona packages.

Quick summary copied from [1]

Percona .deb packages are signed with a key that uses an algorithm now
considered weak. Starting with the next release, Debian and Ubuntu
packages are signed with a new key that uses the much stronger SHA-512
algorithm. All future package release will also contain the new
algorithm.

It’s important that you add the new key before the next release.

See also [2] in trove.

[1] https://www.percona.com/blog/2016/10/13/new-signing-key-for-percona-debian-and-ubuntu-packages/
[2] https://review.openstack.org/#/c/390342/

Change-Id: I4ca5fccf51e51059d4035f7206609c8afa997c94