fix dib for newton
It is now necessary that we explicitly install dib in order to get
things to work in the gate. This is effectively the trove-integration
part of what was commited into master in commit
24dbd37021d71627a0263db22e877d10ee31f526

However, since that is after the refactor of trove-integration into
trove, this change is being made here with a dep from the newton
commit.

Change-Id: Icf24dfc03a1b3549942c5021618e20f93ae8d740